+++ frontend/src/pages/Messages/Conversation.tsx
@@ -1,8 +1,10 @@
 import React from "react";
-import { Grid } from "@mui/material";
+import { Grid, Typography } from "@mui/material";
 import { Messages } from "./Messages";
 import { AddMessageForm } from "./AddMessage";
 import { Socket } from "socket.io-client";
+import { useParams } from "react-router-dom";
+import { MessageDto } from "../../types/message/message.dto";
 
 interface IConversation {
 	socket: Socket | null;
@@ -11,15 +13,32 @@ interface IConversation {
 
 export const Conversation: React.FC<IConversation> = ({ socket, userId }) => {
 
+	const [messages, setMessages] = React.useState<MessageDto[]>([]);
 
+	const params = useParams();
 	React.useEffect(() => {
-		socket?.emit('joinConversation', { friendId: 3})
-	})
+		if (!params.id) {
+			return;
+		}
+		socket?.emit('joinConversation', {friendId: Number(params.id)});
+		socket?.on("messages", (data) => {
+			console.log(data);
+			setMessages([...messages, ...data]);
+		});
+	},[params])
 
 	return (
 		<Grid>
-			<Messages socket={socket} />
-			<AddMessageForm socket={socket} userId={userId} />
+			{!params.id ? <Grid>
+				<Typography>Start chat!</Typography>
+			</Grid>
+			:
+			<Grid>
+				<Messages socket={socket} messages={messages}/>
+				<AddMessageForm socket={socket} userId={userId} />
+			</Grid>
+			}
+
 		</Grid>
 	);
 };